backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
crewai
langchain
langchain-community
langchain-groq
langchain-huggingface
pypdf
faiss-cpu
sentence-transformers
twilio
sqlalchemy
sqlalchemy-json
apscheduler
httpx
